+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Fixed
+- Fixed `NoteableClient.add_cell` callback to look explicity for `new_delta_reply` message type, instead of by transaction_id. This is because if the `new_delta_event` message is received before the `new_delta_reply`, the `add_call` success callback will fail due to schema validation error.
+
 ## Changed
+- Configured `pytest-asyncio` in `pyproject.toml` to set the `asyncio_mode` to "auto".
 - Increase `ClientConfig.ws_timeout` from 10s to 60s
 - Set `open_timeout` on `websockets.connect` in `NoteableClient._connect_rtu_socket`